Consolidate all DefinitionBasedSdJwtVcValidator implementations in a single file.
